renaming kendalls_tau to streaming_correlations
