+++ tensorflow_addons/metrics/__init__.py
@@ -31,4 +31,4 @@ from tensorflow_addons.metrics.multilabel_confusion_matrix import (
 from tensorflow_addons.metrics.r_square import RSquare
 from tensorflow_addons.metrics.geometric_mean import GeometricMean
 from tensorflow_addons.metrics.harmonic_mean import HarmonicMean
-from tensorflow_addons.metrics.kendalls_tau import KendallsTau
+from tensorflow_addons.metrics.streaming_correlations import KendallsTau
